==> tasks/05_association.Rmd <==
---
title: Task. Measures of association
subtitle: Data and statistics with R <br> Research Methods <br><br>
author: Jüri Lillemets
date: "`r Sys.Date()`"
output:
  html_document:
    theme: yeti
    code_folding: hide
editor_options: 
  chunk_output_type: console
---
  
```{r setup, include = F}
# Settings
knitr::opts_chunk$set(include = T, eval = T, echo = T, message = T, warning = T, error = T, 
                      root.dir = '/home/jrl/rcrs')
```

# Categorical variables

## Load and understand data

For the tasks involving categorical variables use the data set *blackman_wheat.csv* in Moodle. The data set includes yield for conventional and semi-dwarf wheat varieties at 7 locations with low/high fertilizer levels. Download the data set to your computer, read it using `read.csv()` and then store it as a data frame.

```{r, echo = F}
wheat <- agridat::blackman.wheat
```

```{r, eval = F}
wheat <- read.csv('blackman_wheat.csv')
```

Next, use `str()` for an overview of the data frame.

```{r}
str(wheat)
```

The variables are described as follows:

- `gen`, genotype
- `loc`, location
- `nitro`, nitrogen fertilizer, low/high
- `yield`, yield (g/m^2)
- `type`, type factor, conventional/semi-dwarf

## Research problem

```{r}
wheat$h.nitro <- wheat$nitro == 'H'
wheat$h.yield <- wheat$yield > mean(wheat$yield)
```

```{r}
conTab <- table(wheat$h.nitro, wheat$h.yield, dnn = c("Nitrogen", "Yield"))
conTab
```

## Odds

Odds of having high yield with high nitrogen vs having high yield without high nitrogen.

```{r}
sum(wheat$h.yield[wheat$h.nitro]) / sum(wheat$h.yield[!wheat$h.nitro])
```

## Risks

Probability of having high yield with high nitrogen

```{r}
sum(wheat$h.yield[wheat$h.nitro]) / sum(wheat$h.yield)
```

# Continuous variables

## Load and understand data

To solve the following exercises use data set *engelstad_nitro.csv* available in Moodle. It contains corn yield response to nitrogen fertilizer for a single variety of corn at two locations over five years. Store this data in R workspace as you did with previous data set.

```{r, echo = F}
corn <- agridat::engelstad.nitro
```

```{r, eval = F}
corn <- read.csv('engelstad_nitro.csv')
```

Again, start by taking a look the data set with `str()`.

```{r}
str(corn)
```

Below is a brief description of the variables.

- `loc`, location, 2 levels
- `year`, year, 1962-1966
- `nitro`, nitrogen fertilizer kg/ha
- `yield`, yield, quintals/ha

## Research problem

```{r}
plot(corn$nitro, corn$yield)
```

Actually discrete, so boxplot...

``` {r}
boxplot(yield ~ nitro, corn)
```

## Correlation

```{r}
cor(corn$nitro, corn$yield, method = 'pearson')
cor(corn$nitro, corn$yield, method = 'spearman')
```

> Which should we prefer, Pearson's or Spearman coeficient of correlation?

## Heatmap

```{r}
corTab <- cor(corn[, c('year', 'nitro', 'yield')], method = 'spearman')
heatmap(corTab)
```
